Basic renaming in python and shell files
