+++ oscovida/metadata.py
@@ -5,7 +5,7 @@ import os
 import pandas as pd
 
 
-
+# FIXME: where should this come from?
 MetadataStorageLocation = os.path.join("coronavirus-metadata", "regions")
